+++ CHANGELOG.md
@@ -8,7 +8,13 @@ and this project adheres to
 
 ## [Unreleased]
 
+## Added
+
 - 📝(doc) add publiccode.yml
+
+## Fixed
+
+- 🐛(back) allow only images to be used with the cors-proxy #781
 
 ## [2.5.0] - 2025-03-18